refactoring

+++ backend/src/Files/SachkovTech.Files.Application/Commands/UploadFiles/UploadFilesHandler.cs
@@ -29,79 +29,90 @@ namespace SachkovTech.Files.Application.Commands.UploadFiles
             CancellationToken cancellationToken = default)
         {
             List<FileId> fileIds = [];
-            List<Error> errors = [];
 
             var uploadAsyncEnum = _fileProvider.UploadFiles(command.Files, cancellationToken);
 
             await foreach (var uploadFileResult in uploadAsyncEnum)
             {
-                if (uploadFileResult.IsSuccess)
-                {
-                    var uploadResult = uploadFileResult.Value;
+                if (!uploadFileResult.IsSuccess) continue;
 
-                    var saveFileResult = await SaveFile(uploadResult, command, cancellationToken);
+                var uploadResult = uploadFileResult.Value;
 
-                    if (saveFileResult.IsSuccess)
-                        fileIds.Add(saveFileResult.Value);
-                    else
-                        errors.Add(saveFileResult.Error);
-                }
-                else
-                {
-                    errors.Add(uploadFileResult.Error);
-                }
+                var saveFileResult = await SaveFile(uploadResult, command, cancellationToken);
+
+                if (saveFileResult.IsSuccess)
+                    fileIds.Add(saveFileResult.Value);
             }
 
             if (fileIds.Count <= 0)
                 return new ErrorList([Error.Failure("file.upload", "Fail to upload files in minio")]);
 
-            var result = new UploadFilesResponse(fileIds, fileIds.Count + errors.Count, errors.Count);
+            var notUploadedFilesCount = command.Files.Count() - fileIds.Count;
+            var result = new UploadFilesResponse(fileIds, fileIds.Count, notUploadedFilesCount);
+
+            _logger.LogInformation("Files uploaded: {notUploadedFilesCount}", notUploadedFilesCount);
 
             return result;
         }
 
 
-        private async Task<Result<FileId, Error>> SaveFile(
+        private async Task<Result<FileId, ErrorList>> SaveFile(
             UploadFilesResult uploadFileResult,
             UploadFilesCommand command,
             CancellationToken cancellationToken)
         {
             var fileData = CreateFileData(uploadFileResult, command);
+            if (fileData.IsFailure)
+                return fileData.Error;
 
-            var saveFileResult = await _filesRepository.Add(fileData, cancellationToken);
+            var saveFileResult = await _filesRepository.Add(fileData.Value, cancellationToken);
 
             if (saveFileResult.IsSuccess)
-                return fileData.Id;
+                return fileData.Value.Id;
 
             var fileLocation = new FileLocation(uploadFileResult.BucketName, uploadFileResult.FilePath);
 
             await _fileProvider.RemoveFile(fileLocation, cancellationToken);
 
-            return saveFileResult.Error;
+            return saveFileResult.Error.ToErrorList();
         }
 
-        private FileData CreateFileData(
+        private Result<FileData, ErrorList> CreateFileData(
             UploadFilesResult uploadFileResult,
             UploadFilesCommand command)
         {
             var fileId = FileId.NewFileId();
 
-            var fileName = FileName.Create(uploadFileResult.FileName).Value;
-            var fileSize = FileSize.Create(uploadFileResult.FileSize).Value;
-            var mimeType = MimeType.Parse(uploadFileResult.FileName).Value;
-            var fileType = FileType.Parse(uploadFileResult.FileName).Value;
-            var ownerType = OwnerType.Create(command.OwnerTypeName).Value;
+            var fileNameResult = FileName.Create(uploadFileResult.FileName);
+            if (fileNameResult.IsFailure)
+                return fileNameResult.Error.ToErrorList();
+
+            var fileSizeResult = FileSize.Create(uploadFileResult.FileSize);
+            if (fileSizeResult.IsFailure)
+                return fileSizeResult.Error.ToErrorList();
+
+            var mimeTypeResult = MimeType.Parse(uploadFileResult.FileName);
+            if (mimeTypeResult.IsFailure)
+                return mimeTypeResult.Error.ToErrorList();
+
+            var fileTypeResult = FileType.Create(uploadFileResult.FileName);
+            if (fileTypeResult.IsFailure)
+                return fileTypeResult.Error.ToErrorList();
+
+            var ownerTypeResult = OwnerType.Create(command.OwnerTypeName);
+            if (ownerTypeResult.IsFailure)
+                return ownerTypeResult.Error.ToErrorList();
 
             return new FileData(
                 fileId,
-                fileName,
+                fileNameResult.Value,
                 command.OwnerId,
                 uploadFileResult.FilePath,
                 true,
-                fileSize,
-                mimeType,
-                fileType,
-                ownerType);
+                fileSizeResult.Value,
+                mimeTypeResult.Value,
+                fileTypeResult.Value,
+                ownerTypeResult.Value);
         }
     }
 }
